Made some very important changes to SignableXMLEnvelope:
Now the envelope may be in a "signed" or "unsigned" state.

By default an envelope is in an unsigned state. Calling
sign() will permanently change the envelope to a signed
state.

When an envelope is in an unsigned state any changes made
to its xml serializer contents will be immediately
reflected when converting the envelope to xml.

When an envelope is in a signed state any changes made
to its xml serializer contents will only be reflected
upon calling sign().

Made Base64Coder work with encoding null data -- it
results in a blank string.

Made cryptor work with signing null strings, it signs
a blank string.