Various changes to dbcore wizard code.
Wizard has been slightly redesigned so that it
should not need to be extended very often --
so it is no longer abstract.

A WizardBuilder interface has been created to
build custom wizards. When this interface is
implemented, it allows a custom task, pool of pages,
page selector, and wizard to be created.

The Wizard and WizardPageNavigator have been
simplified by deferring first/final page determination
to the WizardPageSelector. The Wizard now uses
a passed WizardPagePool to work off of rather than
constructing its own.