Added code to reset read position and read offset to 0 in ManagedInputStream if they meet each other in the middle of the read buffer. This fixes a bug where the "last" boolean parameter for input stream processor's processData() method was not getting set even though the chunk of data was the last chunk in the stream.
Fixed various bugs in StreamCryptor and changed trailer format.

Now the trailer contains just the size of the unencrypted data
as this is all that is necessary (works just like gzip). Fixed
bug that caused the final block to be padded improperly (the
data buffer size was not set properly during decryption).