Set target dyno type when merging.
If merging a map or array on top of another dynmaic object it is now
assumed you want to force the target to the source type.  This is in
line with how basic types are handled.  If this is not done then an
empty source array or map can leave the target as a dyno default empty
string.