Fixed potential shutdown problems with ConnectionService by reverting some code.
Code currently relies on the use of an Operation for servicing an individual call.
If we want to change this to use a Fiber/Actor, then db::actor will need to be
implemented first -- along with libev for IO detection. We might want IO to happen
on a separate Operation where Actors have their states updated when buffers are
ready to be read from. The underlying IO model for Connections/Sockets would need
to change where there is an application layer buffer to read/write to for each
one -- that is then sent out/received as libev (or other backend) indicates.