Added the ability to create a persistent MessageDigest.
Often the value of message digests are checked repeatedly--
as more data is aggregated--but openssl's implementation
finalizes the digest context when you get the value of it,
making it unusable. A flag has been added so that, if
desired, a persistent message digest can be created that
will copy the context before getting its value, so it
can continue to be appended to.