Added feature to EventDaemon to allow scheduled events to be shared.
Now when adding an event to the daemon, a reference parameter can
be specified. If 0 a new event will be created even if it duplicates
another one. If > 0, if the same event on the same interval already
exists, its reference count will be updated by the passed amount and
its count will be increased by the passed amount.