Use ops param for # off threads in thread test.
Switching to non-intuitive use of op count rather than thread count to
specify how many threads are used.  This is done due to the one-to-one
mapping of op count to number of threads in the "threads" test.  This
also will help to normalize results with the "fibers" and "modest"
tests.