Use linking for top-level test.
Top-level test was build using C++ source file includes.  This is a bad
idea in general but specifically when optimization is turned on the
massive generated source file took an unreasonable amount of time and
memory to compile.  This new linking method is MUCH faster.  It however
does use GCC "weak" attribute to mark individual test file main()s as
weak.  And it relies on the linker to toss out individual main()s in
favor of the top-level main which is not marked weak.  This may need to
be ported to other platforms as needed or the top level file could be
just disabled in favor of a test script.