Fixed bug in remove() where rval wasn't being set correctly.
Since the function was defaulting to false and an exception is
only set when the file remove fails and the file does exist,
it was returning false if the file didn't exist and couldn't
be removed (for obvious reasons) ... with no exception set.
The function defaults to true now, only being set to false
if the file remove fails and the file exists.