Updated build system to single platform mode.
Removing the somewhat broken multi-platform build system in favor of a
system that will just build for a single platform type at a time. This
allows removal of a significant amount of duplicated build
infrastructure and will allow for shared use of the build system by
other related projects (ongoing work).