Added AppPlugin call to get wait events.
The App was trying to get wait events from the boot app config. The
events need to be added to the waiter before the sub-app is started. To
do this properly a new method is added to AppPlugins to get a list of
wait events. This can be called before the sub-app is started but is
currently just a static list.