Added delayed logging support.
In order to support outputting to a file that may be specified using
config variables that are not known until after logging is start, a
delayed logging feature is needed. It starts a small in-memory logger
for the bootstrap app and will set the file to use in non-bootstrap
mode. This requires additional support for direct access to the
sub-app parent FileLogger.